partial

+++ OrganizationProfile/StudentInformationClass.cs
@@ -11,12 +11,10 @@ namespace OrganizationProfile
     public class StudentInformationClass
     {
         public static int SetStudentNo = 0;
-        public static int SetContactNo = 0;
+        public static long SetContactNo = 0;
         public static string SetProgram = "";
         public static string SetGender = "";
         public static string SetBirthday = "";
         public static string SetFullName = "";
     }
-
-    
 }
